+++ src/main/java/com/giraone/imaging/video/VideoProvider.java
@@ -1,0 +1,9 @@
+package com.giraone.imaging.video;
+
+import com.giraone.imaging.ThumbnailProvider;
+
+/**
+ *  Interface for thumbnail generation operations on videos (MP4).
+ */
+public interface VideoProvider extends ThumbnailProvider {
+}